+++ HotelBookAPI/Application/HotelBookAPI.Application/Features/RemoveServices/RemoveServicesResponse.cs
@@ -222,7 +222,7 @@ namespace HotelBookAPI.Application.Features.RemoveServices
         public string AdditionalCode2 { get; set; }
         public string AdditionalCode3 { get; set; }
         public string AdditionalCode4 { get; set; }
-        public int AgencyDiscount { get; set; }
+        public double AgencyDiscount { get; set; }
         public bool HasAvailablePromotionCode { get; set; }
 
     }
